+++ src/professions/admin/interfaces/index-template.interface.ts
@@ -1,5 +1,6 @@
 import { CheckboxItems } from 'src/common/interfaces/checkbox-items.interface';
 import { Table } from '../../../common/interfaces/table';
+import { RegulationType } from '../../profession-version.entity';
 import { ProfessionsPresenterView } from '../professions.presenter';
 
 export interface IndexTemplate {
@@ -14,11 +15,13 @@ export interface IndexTemplate {
     nations: string[];
     organisations: string[];
     industries: string[];
+    regulationTypes: RegulationType[];
     changedBy: string[];
   };
 
   nationsCheckboxItems: CheckboxItems[];
   organisationsCheckboxItems: CheckboxItems[];
   industriesCheckboxItems: CheckboxItems[];
+  regulationTypesCheckboxItems: CheckboxItems[];
   changedByCheckboxItems: CheckboxItems[];
 }