+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Delete previously indexed profession versions from Opensearch when publishing
 - Delete profession versions from Opensearch when archiving
 - Search profession titles using Opensearch
+- Allow central users to filter Profession and Organisation listings by regulation type
 
 ## [release-009] - 2022-03-11